-update PlayerComp Edit 부분 (Player Data, Key Binding) 업데이트 PlayerComp 중복 생성 가능하도록 변경 (2명 이상 플레이어를 위한) 일부 경고 처리 수정 BaseComponent에서 Init, Exit 함수 추가(미완)

+++ AmuEngine/AmuEngine/Sources/Map/Map.cpp
@@ -2,11 +2,11 @@
 #include "../Data/Data.h"
 #include "MapManager.h"
 
-Map::Map(std::string mapName_, float width_, float height_, float spacing_) : mapName(mapName_), width(width_), height(height_), spacing(spacing_)
+Map::Map(std::string mapName_, int width_, int height_, float spacing_) : mapName(mapName_), width(width_), height(height_), spacing(spacing_)
 {
     // 맵 크기에 맞게 빈 타일로 초기화
     isViewGrid = true;
-    tiles = std::vector<std::vector<Tile>>(height, std::vector<Tile>(width, Tile(Tile::TileType::Empty, false, 0, 0)));
+    tiles = std::vector<std::vector<Tile>>((unsigned int)height, std::vector<Tile>((unsigned int)width, Tile(Tile::TileType::Empty, false, 0, 0)));
     grid = Grid((float)width, (float)spacing);
     MapManager::GetInstance().AddMap(this);
 }
@@ -15,11 +15,11 @@ void Map::addItem(Item item) {
     items.push_back(item);
 }
 
-void Map::addSpawnPoint(float x, float y) {
+void Map::addSpawnPoint(int x, int y) {
     spawnPoints.push_back({ x, y });
 }
 
-void Map::setTile(float x, float y, Tile::TileType type, bool destructible) {
+void Map::setTile(int x, int y, Tile::TileType type, bool destructible) {
     if (x >= 0 && x < width && y >= 0 && y < height) {
         tiles[y][x] = Tile(type, destructible, x, y);
     }
